+++ src/main/java/com/fo/test/task/repository/AuthorRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.fo.test.task.model.Author;
 
 @Repository
-public interface AuthorRepository extends JpaRepository<Author, Integer>{
+public interface AuthorRepository extends JpaRepository<Author, Long>{
 
 }
