+++ src/main/java/com/fo/test/task/repository/Book_AuthorRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.fo.test.task.model.Book_Author;
 
 @Repository
-public interface Book_AuthorRepository extends JpaRepository<Book_Author, Integer>{
+public interface Book_AuthorRepository extends JpaRepository<Book_Author, Long>{
 
 }
